+++ src/oldas/__init__.py
@@ -16,13 +16,13 @@ __licence__ = "MIT"
 
 ##############################################################################
 # Local imports.
-from ._prefixes import Prefix, id_is_a_feed, id_is_a_folder, id_is_an_article
-from ._states import State
 from .articles import Article, Articles
 from .exceptions import OldASError, OldASInvalidLogin
 from .folders import Folder, Folders
 from .id_streams import ArticleID, ArticleIDs
+from .prefixes import Prefix, id_is_a_feed, id_is_a_folder, id_is_an_article
 from .session import Session
+from .states import State
 from .subscriptions import Subscription, Subscriptions
 from .unread import Count, Counts, Unread