+++ src/oldas/articles.py
@@ -11,12 +11,12 @@ from typing import Any, AsyncIterator, Iterable, Literal, NamedTuple
 
 ##############################################################################
 # Local imports.
-from ._prefixes import id_is_a_folder
-from ._states import State
-from ._types import OldList, RawData
 from .folders import Folder
+from .prefixes import id_is_a_folder
 from .session import Session
+from .states import State
 from .subscriptions import Subscription
+from .types import OldList, RawData
 
 ##############################################################################
 Direction = Literal["ltr", "rtl"]