+++ src/oldas/id_streams.py
@@ -7,10 +7,10 @@ from typing import Any, AsyncIterator, NamedTuple
 
 ##############################################################################
 # Local imports.
-from ._prefixes import Prefix
-from ._states import State
-from ._types import OldList, RawData
+from .prefixes import Prefix
 from .session import Session
+from .states import State
+from .types import OldList, RawData
 
 
 ##############################################################################