+++ src/oldas/session.py
@@ -11,9 +11,9 @@ from httpx import AsyncClient, HTTPStatusError, ReadTimeout, RequestError, Respo
 ##############################################################################
 # Local imports.
 from . import __version__
-from ._states import State
-from ._types import RawData
 from .exceptions import OldASError, OldASInvalidLogin, OldASLoginNeeded
+from .states import State
+from .types import RawData
 
 
 ##############################################################################