+++ src/oldas/unread.py
@@ -11,9 +11,9 @@ from typing import NamedTuple
 
 ##############################################################################
 # Local imports.
-from ._prefixes import Prefix, id_is_a
-from ._types import OldList, RawData
+from .prefixes import Prefix, id_is_a
 from .session import Session
+from .types import OldList, RawData
 
 
 ##############################################################################